requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.12.0
django-formtools==2.3
django-otp==1.0.6
django-phonenumber-field==5.2.0
django-two-factor-auth==1.13.1
djangorestframework==3.12.4
phonenumbers==8.12.27
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
qrcode==6.1
six==1.16.0
sqlparse==0.4.1
